+++ build-env/assets/changelog.md
@@ -1,5 +1,10 @@
+## v1.12 beta
+
+- **PluginLoader**: Errors in plugins are now properly reported
+- **PluginLoader**: Fixed pug doctype
+- **CORE**: CardManager now use plugins' profile data to determine binded attribute
+- **WebUI**: Edit button will not appear if no custom profile pages exists
+
 ## v1.10 beta
 
 - First WebUI public beta
-- **Note**: API has gone through a complete overhaul compared to the first beta release.
-  Please make sure to follow the new plugin template and the documentation.